use InvocationInterceptor based approach
which requires additionally some store to save extracted code
and examples (use object instead of junit store for simplicity)
Write pseudo test which updates README and always runs last
(use a custom ClassOrderer which puts this test last)

moreover:
- move sources from main to test
- remove Spec from file name (use ..Examples instead)
- check that we don't define a snippet twice (fix ThirdPartyExample)
- move documentation to ReadmeTest only
