+++ gradle/build-logic/root-build/src/main/kotlin/build-logic.root-build.gradle.kts
@@ -93,7 +93,6 @@ apiValidation {
             "atrium-specs",
             "atrium-verbs-internal",
             "readme-examples",
-            "readme-examples-junit5"
         )
     )
     val kotlinVersion = KotlinVersion.fromVersion(buildParameters.kotlin.version)
