--- misc/tools/readme-examples-junit5/src/main/resources/META-INF/services/org.junit.platform.engine.TestEngine
@@ -1,1 +1,0 @@
-readme.examples.JUnitReadmeTestEngine
