+++ misc/tools/readme-examples/src/test/resources/junit-platform.properties
@@ -1,0 +1,1 @@
+junit.jupiter.testclass.order.default=readme.examples.jupiter.ReadmeTestClassOrderer
